(Just accidentally pushed to main branch. The changes will now be reviewable.
GUI to display the encounterview for merchant, enemy, and the final boss, where an additional window is triggered when the final boss is slain.

GamePanel package has been removed as well.

TODO (as shown within the code):
will be complete after incorporating the features from the combat loop and some feature for switching the state.